wrong tooltip on method down arrow, Ctrl-shift-I on directory NPE

+++ source/com/intellij/openapi/vfs/impl/local/VirtualFileImpl.java
@@ -307,7 +307,7 @@ public class VirtualFileImpl extends VirtualFile {
     final ProvidedContent content = getProvidedContent();
     InputStream in = content.getInputStream();
     try {
-      return FileUtil.loadBytes(in, (int)content.getLength());
+      return FileUtil.loadBytes(in, content.getLength());
     }
     finally {
       in.close();
